1.Finish custom history AlertDialog. 2.Begin adding missing buttons to activity_calculator.xml
Added a new icon for the history dialog (ic_history6) and added a custom adapter that places the history data in the list. Also, an inner class (HistoryItem) was made to store the history since we are saving both the expression and it's result.

Added some of the missing buttons to the XML and in Calculator.java I added their member variables, assigned them their view, and set the onClick listener. Now the only thing missing is to add their functionality on the listener.